% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crmn-package.R
\docType{package}
\name{crmn}
\alias{crmn}
\alias{crmn-package}
\title{CRMN}
\description{
Normalize metabolomics data using CCMN and other methods
}
\details{
\tabular{ll}{
Package: \tab crmn \cr
Type: \tab Package \cr
Developed since: \tab 2009-05-14 \cr
Depends: \tab Biobase, pcaMethods (>= 1.20.2), pls, methods \cr
License: \tab GPL (>=3) \cr
LazyLoad: \tab yes \cr
}

A package implementing the 'Cross-contribution compensating
multiple standard normalization' described in Redestig et al. (2009)
Analytical Chemistry, https://doi.org/10.1021/ac901143w. Can be used to
normalize metabolomics data. Do \code{openVignette("crmn")} to see
the manual.
}
\author{
Henning Redestig
}
