% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{method-methods}
\alias{method}
\alias{method,nFit-method}
\alias{method-methods}
\title{Accessor for the method}
\usage{
method(object, ...)

method(object, ...)
}
\arguments{
\item{object}{an \code{nFit} object}

\item{...}{not used}
}
\value{
the method (content differs between normlization methods)
}
\description{
Get the method
}
\author{
Henning Redestig
}

