\name{method}
\title{Accessor for the method}
\usage{method(object, ...)}
\description{Get the method}
\value{the method (content differs between normlization methods)}
\author{Henning Redestig \email{henning@psc.riken.jp}}
\alias{method}
\alias{method,nFit-method}
\arguments{\item{object}{an \code{nFit} object}
\item{...}{not used}}
