% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R
\name{mexprs-methods}
\alias{mexprs-methods}
\alias{mexprs}
\alias{mexprs,matrix-method}
\alias{mexprs,ExpressionSet-method}
\title{Matrix safe accessor of expression slot}
\usage{
mexprs(object)

mexprs(object)

\S4method{mexprs}{ExpressionSet}(object)
}
\arguments{
\item{object}{an \code{ExpressionSet} or \code{matrix}}
}
\value{
the expression data
}
\description{
Get the expression data from an \code{ExpressionSet} or
just return the given matrix
}
\examples{
data(mix)
head(mexprs(mix))
head(mexprs(exprs(mix)))
}
\author{
Henning Redestig
}
