% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R
\name{mexprs-rep-methods}
\alias{mexprs-rep-methods}
\alias{mexprs<-}
\alias{mexprs<-,ExpressionSet,matrix-method}
\alias{mexprs<-,matrix,matrix-method}
\title{Accessor}
\usage{
mexprs(object) <- value

\S4method{mexprs}{ExpressionSet,matrix}(object) <- value

mexprs(object) <- value
}
\arguments{
\item{object}{an \code{ExpressionSet} or \code{matrix}}

\item{value}{the value to assign}
}
\value{
the expression data
}
\description{
Matrix safe setter of expression slot
}
\details{
Set the expression data in an \code{ExpressionSet} or
just return the given matrix
}
\examples{
data(mix)
test <- mix
mexprs(test) <- exprs(mix) * 0
head(mexprs(test))
test <- exprs(mix)
mexprs(test) <- test * 0
head(mexprs(test))
}
\author{
Henning Redestig
}
