\name{model}
\title{Accessor for the model...}
\usage{model(object, ...)}
\description{Accessor for the model}
\details{Get the model}
\value{the model (content differs between normlization models)}
\author{Henning Redestig \email{henning@psc.riken.jp}}
\alias{model}
\alias{model,nFit-method}
\arguments{\item{object}{an \code{nFit} object}
\item{...}{not used}}
