% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm.R
\name{pcaMuffle}
\alias{pcaMuffle}
\title{Muffle the pca function}
\usage{
pcaMuffle(w)
}
\arguments{
\item{w}{a warning}
}
\value{
nothing
}
\description{
PCA and Q2 issues warnings about biasedness and poorly estimated
PCs. The first is non-informative and the poorly estimated PCs will
show up as poor overfitting which leads to a choice of fewer PCs
i.e. not a problem. This function is mean to muffle those warnings.
Only used for version of pcaMethods before 1.26.0.
}
\author{
Henning Redestig
}
