% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{sFit-method}
\alias{sFit}
\alias{sFit,nFit-method}
\alias{sFit-method}
\title{Accessor for the standards model}
\usage{
sFit(object, ...)

sFit(object, ...)
}
\arguments{
\item{object}{an \code{nFit} object}

\item{...}{not used}
}
\value{
the sFit is only defined for CRMN
}
\description{
Get the sFit
}
\author{
Henning Redestig
}

