\name{sFit}
\title{Accessor for the standards model}
\usage{sFit(object, ...)}
\description{Get the sFit}
\value{the sFit is only defined for CRMN}
\author{Henning Redestig \email{henning@psc.riken.jp}}
\alias{sFit}
\alias{sFit,nFit-method}
\arguments{\item{object}{an \code{nFit} object}
\item{...}{not used}}
