% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{show_nfit}
\alias{show_nfit}
\title{Show nfit}
\usage{
show_nfit(object)
}
\arguments{
\item{object}{the \code{nFit} object}
}
\value{
prints some basic information
}
\description{
Show method for nFit
}
\author{
Henning Redestig
}
