% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{standards_eset}
\alias{standards_eset}
\alias{standards,ExpressionSet,missing-method}
\title{Accessor for the Internal Standards}
\usage{
standards_eset(object, where = "tag", what = "IS", ...)
}
\arguments{
\item{object}{an \code{ExpressionSet}}

\item{where}{Column index or name in fData which equals
\code{what} for the ISs}

\item{what}{What the column \code{where} equals for ISs}

\item{...}{not used}
}
\value{
subsetted dataset
}
\description{
Subset an data set to only contain the labeled internal standards.
}
\examples{
data(mix)
standards(mix)
fData(mix)$test <- fData(mix)$tag
standards(mix, where="test")
}
\author{
Henning Redestig
}
