% Generated by roxygen2 (4.0.2): do not edit by hand
\name{analytes_other}
\alias{analytes,data.frame,logical-method}
\alias{analytes,matrix,logical-method}
\alias{analytes_other}
\title{Accessor for the analytes}
\usage{
analytes_other(object, standards, ...)
}
\arguments{
\item{object}{an \code{ExpressionSet}}

\item{standards}{a logical vector indicating which rows are
internal standards}

\item{...}{not used}
}
\value{
\code{ExpressionSet}
}
\description{
Subset an expression set to remove the internal standards
}
\examples{
data(mix)
analytes(exprs(mix), fData(mix)$tag == 'IS')
}
\author{
Henning Redestig
}

