\name{crmn}
\title{CRMN}
\description{Normalize metabolomics data using the CRMN method}
\details{\tabular{ll}{
Package: \tab metrik \cr
Type: \tab Package \cr
Version: \tab 0.0.3 \cr
Date: \tab 2009-05-14 \cr
Depends: \tab Biobase, pcaMethods (>= 1.20.2), pls, methods \cr
License: \tab GPL (>=3) \cr
LazyLoad: \tab yes \cr
}

A package implementing the 'Cross-contribution robust 
normalization using multiple internal standards'. Can be used to
normalize metabolomics data. Do \code{openVignette("crmn")} to see
the manual.}
\alias{crmn}
\docType{package}
\author{Henning Redestig \email{henning@psc.riken.jp}}
