\name{mix}
\title{Dilution mixture dataset.}
\usage{data(mix)}
\description{Mixture dilution series}
\details{Multi-component dilution series. GC-TOF/MS measurements by Miyako Kusano.
Input concentrations are known and given in the original publication.}
\alias{mix}
\docType{data}
\author{Henning Redestig \email{henning@psc.riken.jp}}
\examples{data(mix)
fData(mix)
exprs(mix)
pData(mix)}
