\docType{data}
\name{mix}
\alias{mix}
\title{Dilution mixture dataset.}
\description{
  Mixture dilution series
}
\details{
  Multi-component dilution series. GC-TOF/MS measurements
  by Miyako Kusano. Input concentrations are known and
  given in the original publication.
}
\examples{
data(mix)
 fData(mix)
 exprs(mix)
 pData(mix)
}
\author{
  Henning Redestig
}
\keyword{datasets}

