% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crmn-package.R
\docType{data}
\name{mix}
\alias{mix}
\title{Dilution mixture dataset.}
\usage{
data(mix)
}
\description{
Mixture dilution series
}
\details{
Multi-component dilution series. GC-TOF/MS measurements by Miyako Kusano.
Input concentrations are known and given in the original publication.
}
\examples{
 data(mix)
 fData(mix)
 exprs(mix)
 pData(mix)
}
\author{
Henning Redestig
}
