% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{model-methods}
\alias{model}
\alias{model,nFit-method}
\alias{model-methods}
\title{Accessor for the model}
\usage{
model(object, ...)

model(object, ...)
}
\arguments{
\item{object}{an \code{nFit} object}

\item{...}{not used}
}
\value{
the model (content differs between normlization models)
}
\description{
Get the model
}
\author{
Henning Redestig
}

