% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R
\name{sFit-method}
\alias{sFit-method}
\alias{sFit}
\alias{sFit,nFit-method}
\title{Accessor for the standards model}
\usage{
sFit(object, ...)

sFit(object, ...)
}
\arguments{
\item{object}{an \code{nFit} object}

\item{...}{not used}
}
\value{
the sFit is only defined for CRMN
}
\description{
Get the sFit
}
\author{
Henning Redestig
}
