% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R
\name{show}
\alias{show}
\alias{show,nFit-method}
\title{Show method for nFit}
\usage{
\S4method{show}{nFit}(object)
}
\arguments{
\item{object}{the \code{nFit} object}
}
\value{
prints some basic information
}
\description{
Show some basic information for an nFit model
}
\examples{
data(mix)
normFit(mix, "avg")
}
\author{
Henning Redestig
}
