\name{show}
\alias{show}
\alias{show,nFit-method}
\title{Show method for nFit}
\arguments{
  \item{object}{the \code{nFit} object}
}
\value{
  prints some basic information
}
\description{
  Show some basic information for an nFit model
}
\examples{
data(mix)
normFit(mix, "avg")
}
\author{
  Henning Redestig
}

