\name{standards_eset}
\title{Accessor for the Internal Standards...}
\usage{standards_eset(object, ...)}
\description{Accessor for the Internal Standards}
\details{Subset an data set to only contain the labeled internal standards.}
\value{subsetted dataset}
\author{Henning Redestig \email{henning@psc.riken.jp}}
\alias{standards_eset}
\alias{standards,ExpressionSet,missing-method}
\arguments{\item{object}{an \code{ExpressionSet}}
\item{...}{not used}}
\examples{data(mix)
standards(mix)}
